Refactoring: moved all testfile related code to one file
