+++ src/lib/createTestFile.js
@@ -1,7 +1,8 @@
 "use strict";
 var webpack = require("./webpackPromise"),
     handleWebpackResult = require("./handleWebpackResult"),
-    p = require("path");
+    p = require("path"),
+    stripTestExtensions = require("./testFileHelpers").stripTestExtensions;
 
 function createEntryModules(karmaPath, srcPath, filenames) {
     var entryModules = {};
@@ -10,7 +11,7 @@ function createEntryModules(karmaPath, srcPath, filenames) {
     }
 
     filenames.forEach(function(file) {
-        var featurename = require("path").relative(srcPath, file.substr(0, file.length - 8)).replace(/\//g, "_");
+        var featurename = require("path").relative(srcPath, stripTestExtensions(file)).replace(/\//g, "_");
         entryModules[p.join(karmaPath, featurename)] = file;
         console.log("    * " + featurename + " (" + file + ")." );
     });
